backend-python/requirements.txt
Django==2.0.1
django-cors-headers==2.1.0
djangorestframework==3.7.7
pytz==2017.3
redis==2.10.6
